Fix `nodes_up`, `last_layer_nodes` bug
Only `input_layer` was being overwritten, so there are 10 nodes, but in
later steps, `nodes_up` and `last_layer_nodes` are used to create
extraneous edges to nodes that nodes that don't exist, corrupting the
output.

The indentation bug in previous commit happened to "fix" this by going
through the graph display code multiple times, with the earlier times
overwriting the variables for the next cycle, so the final output image
happened to be correct, with one bug counteracting the other.